Refactor FCI strategy to use R0R4StrategyTestBased directly
Replaced RecursiveDiscriminatingPathRule with R0R4StrategyTestBased for FCI strategy configuration across all related files. Centralized the special and default configuration methods in R0R4StrategyTestBased to eliminate redundancy and improve maintainability. This simplifies the code and streamlines the orientation logic in FCI algorithms.
