+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/RecursiveDiscriminatingPathRule.java
@@ -1,11 +1,9 @@
 package edu.cmu.tetrad.search;
 
-import edu.cmu.tetrad.data.Knowledge;
 import edu.cmu.tetrad.graph.Endpoint;
 import edu.cmu.tetrad.graph.Graph;
 import edu.cmu.tetrad.graph.GraphUtils;
 import edu.cmu.tetrad.graph.Node;
-import edu.cmu.tetrad.search.test.MsepTest;
 import edu.cmu.tetrad.search.utils.*;
 import edu.cmu.tetrad.util.SublistGenerator;
 
@@ -24,6 +22,60 @@ import java.util.concurrent.ForkJoinPool;
  * @author josephramsey
  */
 public class RecursiveDiscriminatingPathRule {
+
+    /**
+     * Checks for conditional independence between two nodes {@code x} and {@code y} in the context of
+     * a given blocking set, recursively analyzing potential discriminating paths. This method evaluates
+     * combinations of nodes in the blocking set that satisfy independence conditions.
+     *
+     * @param test                 The independence test object used to evaluate conditional independence.
+     * @param x                    The first node for which independence is being checked.
+     * @param y                    The second node for which independence is being checked.
+     * @param blocking             The initial set of nodes considered as the blocking set for independence tests.
+     * @param vNodes               The subset of nodes identifying possible colliders in the analysis.
+     * @param discriminatingPath   The discriminating path data structure containing relevant path information.
+     * @param ensureMarkovHelper   A helper object for ensuring adherence to the Markov property during the checks.
+     * @return                     {@code true} if the nodes {@code x} and {@code y} are conditionally independent
+     *                             given the blocking set; otherwise, {@code false}.
+     * @throws InterruptedException If the process is interrupted during execution.
+     */
+    public static boolean checkIndependenceRecursive(IndependenceTest test, Node x, Node y, Set<Node> blocking, Set<Node> vNodes,
+                                                     DiscriminatingPath discriminatingPath, EnsureMarkov ensureMarkovHelper) throws InterruptedException {
+
+        List<Node> vs = new ArrayList<>();
+        List<Node> nonVs = new ArrayList<>();
+
+        for (Node v : blocking) {
+            if (vNodes.contains(v)) {
+                vs.add(v);
+            } else {
+                nonVs.add(v);
+            }
+        }
+
+        Node v = discriminatingPath.getV();
+        vs.remove(v);
+
+        SublistGenerator generator = new SublistGenerator(vs.size(), vs.size());
+        int[] choice;
+
+        while ((choice = generator.next()) != null) {
+            Set<Node> newBlocking = GraphUtils.asSet(choice, vs);
+            newBlocking.add(v);
+            newBlocking.addAll(nonVs);
+
+            // You didn't condition on any colliders. V is in the set. So V is a noncollider.
+            boolean independent = ensureMarkovHelper != null ? ensureMarkovHelper.markovIndependence(x, y, newBlocking) :
+                    test.checkIndependence(x, y, newBlocking).isIndependent();
+
+            if (independent) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
 
     /**
      * Finds the set of nodes (separator set) for the Recursive Discriminating Path rule in a graph.
@@ -177,115 +229,6 @@ public class RecursiveDiscriminatingPathRule {
                 forkJoinPool.shutdownNow();
             }
         }
-    }
-
-    /**
-     * Provides a special configuration for creating an instance of FciOrientDataExaminationStrategy.
-     *
-     * @param test      the IndependenceTest object used by the strategy
-     * @param knowledge the Knowledge object used by the strategy
-     * @param verbose   boolean indicating whether to provide verbose output
-     * @return a configured FciOrientDataExaminationStrategy object
-     * @throws IllegalArgumentException if test or knowledge is null
-     */
-    public static R0R4Strategy specialConfiguration(IndependenceTest test, Knowledge knowledge, boolean verbose) {
-        if (test == null) {
-            throw new IllegalArgumentException("Test is null.");
-        }
-
-        if (knowledge == null) {
-            throw new IllegalArgumentException("Knowledge is null.");
-        }
-
-        if (test instanceof MsepTest) {
-            R0R4Strategy r0R4Strategy = defaultConfiguration(((MsepTest) test).getGraph(), knowledge);
-            R0R4StrategyTestBased _r0R4Strategy = (R0R4StrategyTestBased) r0R4Strategy;
-            _r0R4Strategy.setVerbose(verbose);
-            return _r0R4Strategy;
-        } else {
-            R0R4StrategyTestBased strategy = new R0R4StrategyTestBased(test);
-            strategy.setKnowledge(knowledge);
-            strategy.setVerbose(verbose);
-            return strategy;
-        }
-    }
-
-    /**
-     * Returns a default configuration of the FciOrientDataExaminationStrategy object.
-     *
-     * @param dag       the graph representation
-     * @param knowledge the Knowledge object used by the strategy
-     * @return a default configured FciOrientDataExaminationStrategy object
-     */
-    public static R0R4Strategy defaultConfiguration(Graph dag, Knowledge knowledge) {
-        return defaultConfiguration(new MsepTest(dag), knowledge);
-    }
-
-    /**
-     * Returns a default configuration of the FciOrientDataExaminationStrategy object.
-     *
-     * @param test      the IndependenceTest object used by the strategy
-     * @param knowledge the Knowledge object used by the strategy
-     * @return a configured FciOrientDataExaminationStrategy object
-     * @throws IllegalArgumentException if test or knowledge is null
-     */
-    public static R0R4Strategy defaultConfiguration(IndependenceTest test, Knowledge knowledge) {
-        R0R4StrategyTestBased strategy = new R0R4StrategyTestBased(test);
-        strategy.setKnowledge(knowledge);
-        return strategy;
-    }
-
-    /**
-     * Checks for conditional independence between two nodes {@code x} and {@code y} in the context of
-     * a given blocking set, recursively analyzing potential discriminating paths. This method evaluates
-     * combinations of nodes in the blocking set that satisfy independence conditions.
-     *
-     * @param test                 The independence test object used to evaluate conditional independence.
-     * @param x                    The first node for which independence is being checked.
-     * @param y                    The second node for which independence is being checked.
-     * @param blocking             The initial set of nodes considered as the blocking set for independence tests.
-     * @param vNodes               The subset of nodes identifying possible colliders in the analysis.
-     * @param discriminatingPath   The discriminating path data structure containing relevant path information.
-     * @param ensureMarkovHelper   A helper object for ensuring adherence to the Markov property during the checks.
-     * @return                     {@code true} if the nodes {@code x} and {@code y} are conditionally independent
-     *                             given the blocking set; otherwise, {@code false}.
-     * @throws InterruptedException If the process is interrupted during execution.
-     */
-    public static boolean checkIndependenceRecursive(IndependenceTest test, Node x, Node y, Set<Node> blocking, Set<Node> vNodes,
-                                                     DiscriminatingPath discriminatingPath, EnsureMarkov ensureMarkovHelper) throws InterruptedException {
-
-        List<Node> vs = new ArrayList<>();
-        List<Node> nonVs = new ArrayList<>();
-
-        for (Node v : blocking) {
-            if (vNodes.contains(v)) {
-                vs.add(v);
-            } else {
-                nonVs.add(v);
-            }
-        }
-
-        Node v = discriminatingPath.getV();
-        vs.remove(v);
-
-        SublistGenerator generator = new SublistGenerator(vs.size(), vs.size());
-        int[] choice;
-
-        while ((choice = generator.next()) != null) {
-            Set<Node> newBlocking = GraphUtils.asSet(choice, vs);
-            newBlocking.add(v);
-            newBlocking.addAll(nonVs);
-
-            // You didn't condition on any colliders. V is in the set. So V is a noncollider.
-            boolean independent = ensureMarkovHelper != null ? ensureMarkovHelper.markovIndependence(x, y, newBlocking) :
-                    test.checkIndependence(x, y, newBlocking).isIndependent();
-
-            if (independent) {
-                return true;
-            }
-        }
-
-        return false;
     }
 
     /**
