+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/SvarGfci.java
@@ -163,7 +163,7 @@ public final class SvarGfci implements IGraphSearch {
 
         modifiedR0(fgesGraph);
 
-        R0R4StrategyTestBased strategy = (R0R4StrategyTestBased) RecursiveDiscriminatingPathRule.specialConfiguration(independenceTest,
+        R0R4StrategyTestBased strategy = (R0R4StrategyTestBased) R0R4StrategyTestBased.specialConfiguration(independenceTest,
                 knowledge, verbose);
         strategy.setDepth(-1);
         strategy.setMaxLength(-1);
